pyproject.toml
[project]
name = "solver-verifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "fastapi-static-files>=0.1.0",
    "markitdown[pdf]>=0.1.2",
    "openai>=1.99.6",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
